move photo layout to common
