+++ src/common/components/PhotoLayout/styled.tsx
