+++ src/common/components/PhotoLayout/types.ts
